Removed most of the noise from part 1.
Removed most of the game code from part 1 of the CDIO project.
Some stuff needs to be decided at some point. Do we want a
GameController or is Game the controller?
TestMatadorRafleBaeger is still there because we can use it to test that
we do not break the Randomness of the dice.